Bugfix: Added hours & minutes processing to the Last Updated metadata
If the difference in datetime between last updated and current time is in hours or minutes,
this update will fix the bug where there will be a error since there wasn't a if statement
which compared the datatime in terms of hours & minutes. Now, the last_updated will take either
hours or minutes as the time.